🍵 remove SJPP preprocessor for `TeaVM` build

+++ src/main/java/net/sourceforge/plantuml/utils/Log.java
@@ -45,76 +45,76 @@ import java.util.function.Supplier;
 import net.sourceforge.plantuml.ProgressBar;
 import net.sourceforge.plantuml.cli.GlobalConfig;
 import net.sourceforge.plantuml.cli.GlobalConfigKey;
+import net.sourceforge.plantuml.teavm.TeaVM;
 
 public abstract class Log {
-	// ::comment when __HAXE__
+
 	private static final long start = System.currentTimeMillis();
 
 	private static String format(String s) {
 		final StringBuilder sb = new StringBuilder();
-		// ::comment when __TEAVM__
-		final long delta = System.currentTimeMillis() - start;
-		// final HealthCheck healthCheck = Performance.getHealthCheck();
-		// final long cpu = healthCheck.jvmCpuTime() / 1000L / 1000L;
-		// final long dot = healthCheck.dotTime().getSum();
+		if (!TeaVM.isTeaVM()) {
+			final long delta = System.currentTimeMillis() - start;
+			// final HealthCheck healthCheck = Performance.getHealthCheck();
+			// final long cpu = healthCheck.jvmCpuTime() / 1000L / 1000L;
+			// final long dot = healthCheck.dotTime().getSum();
 
-		final long freeMemory = Runtime.getRuntime().freeMemory();
-		final long maxMemory = Runtime.getRuntime().maxMemory();
-		final long totalMemory = Runtime.getRuntime().totalMemory();
-		final long usedMemory = totalMemory - freeMemory;
-		final int threadActiveCount = Thread.activeCount();
+			final long freeMemory = Runtime.getRuntime().freeMemory();
+			final long maxMemory = Runtime.getRuntime().maxMemory();
+			final long totalMemory = Runtime.getRuntime().totalMemory();
+			final long usedMemory = totalMemory - freeMemory;
+			final int threadActiveCount = Thread.activeCount();
 
-		sb.append("(");
-		sb.append(delta / 1000L);
-		sb.append(".");
-		sb.append(String.format("%03d", delta % 1000L));
-		// if (cpu != -1) {
-		// sb.append(" - ");
-		// sb.append(cpu / 1000L);
-		// sb.append(".");
-		// sb.append(String.format("%03d", cpu % 1000L));
-		// }
-		// sb.append(" - ");
-		// sb.append(dot / 1000L);
-		// sb.append(".");
-		// sb.append(String.format("%03d", dot % 1000L));
-		// sb.append("(");
-		// sb.append(healthCheck.dotTime().getNb());
-		// sb.append(")");
-		sb.append(" - ");
-		final long total = totalMemory / 1024 / 1024;
-		final long free = freeMemory / 1024 / 1024;
-		sb.append(total);
-		sb.append(" Mo) ");
-		sb.append(free);
-		sb.append(" Mo - ");
-		sb.append(s);
-		// ::done
+			sb.append("(");
+			sb.append(delta / 1000L);
+			sb.append(".");
+			sb.append(String.format("%03d", delta % 1000L));
+			// if (cpu != -1) {
+			// sb.append(" - ");
+			// sb.append(cpu / 1000L);
+			// sb.append(".");
+			// sb.append(String.format("%03d", cpu % 1000L));
+			// }
+			// sb.append(" - ");
+			// sb.append(dot / 1000L);
+			// sb.append(".");
+			// sb.append(String.format("%03d", dot % 1000L));
+			// sb.append("(");
+			// sb.append(healthCheck.dotTime().getNb());
+			// sb.append(")");
+			sb.append(" - ");
+			final long total = totalMemory / 1024 / 1024;
+			final long free = freeMemory / 1024 / 1024;
+			sb.append(total);
+			sb.append(" Mo) ");
+			sb.append(free);
+			sb.append(" Mo - ");
+			sb.append(s);
+		}
 
 		return sb.toString();
-		
+
 	}
 
 	public static void println(Object s) {
 	}
-	// ::done
 
 	public synchronized static void debug(Supplier<String> msgSupplier) {
 	}
 
 	public synchronized static void info(Supplier<String> msgSupplier) {
-		// ::comment when __CORE__ or __HAXE__ or __TEAVM__
-		if (GlobalConfig.getInstance().boolValue(GlobalConfigKey.VERBOSE)) {
-			ProgressBar.clear();
-			System.err.println(format(msgSupplier.get()));
+		if (!TeaVM.isTeaVM()) {
+			if (GlobalConfig.getInstance().boolValue(GlobalConfigKey.VERBOSE)) {
+				ProgressBar.clear();
+				System.err.println(format(msgSupplier.get()));
+			}
 		}
-		// ::done
 	}
 
 	public synchronized static void error(String s) {
-		// ::comment when __CORE__ or __HAXE__ or __TEAVM__
-		ProgressBar.clear();
-		// ::done
+		if (!TeaVM.isTeaVM()) {
+			ProgressBar.clear();
+		}
 		System.err.println(s);
 	}
 
@@ -129,14 +129,15 @@ public abstract class Log {
 	}
 
 	public static void deletePerfLogFile() {
-		// ::comment when __TEAVM__
-		final File file = new File(PERFLOG_FILENAME);
-		if (file.exists()) {
-			if (!file.delete()) {
-				error("Cannot del " + PERFLOG_FILENAME);
+		if (!TeaVM.isTeaVM()) {
+			final File file = new File(PERFLOG_FILENAME);
+			if (file.exists()) {
+				if (!file.delete()) {
+					error("Cannot del " + PERFLOG_FILENAME);
+				}
+
 			}
 		}
-		// ::done
 	}
 
 	private synchronized static void perflogInternal(String line) {
@@ -156,19 +157,19 @@ public abstract class Log {
 	}
 
 	public static synchronized void logStackTrace(String s) {
-		// ::comment when __TEAVM__
-		if (firstCall) {
-			deletePerfLogFile();
-			firstCall = false;
+		if (!TeaVM.isTeaVM()) {
+			if (firstCall) {
+				deletePerfLogFile();
+				firstCall = false;
+			}
+			try (FileWriter fw = new FileWriter(PERFLOG_FILENAME, true); PrintWriter pw = new PrintWriter(fw)) {
+				final Exception e = new Exception("StackTrace " + s);
+				e.fillInStackTrace();
+				e.printStackTrace(pw);
+			} catch (IOException ex) {
+				ex.printStackTrace();
+			}
 		}
-		try (FileWriter fw = new FileWriter(PERFLOG_FILENAME, true); PrintWriter pw = new PrintWriter(fw)) {
-			final Exception e = new Exception("StackTrace " + s);
-			e.fillInStackTrace();
-			e.printStackTrace(pw);
-		} catch (IOException ex) {
-			ex.printStackTrace();
-		}
-		// ::done
 	}
 
 }
